Extract manual send handlers into dedicated package
