+++ emailbot/handlers/__init__.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+from __future__ import annotations
+
+from .manual_send import (
+    start,
+    manual_mode,
+    select_group,
+    proceed_to_group,
+    send_all,
+)
+
+__all__ = [
+    "start",
+    "manual_mode",
+    "select_group",
+    "proceed_to_group",
+    "send_all",
+]
