+++ emailbot/handlers/history.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from __future__ import annotations
+
+
